early kartverket import

+++ converter/src/main/kotlin/no/entur/netex_photon/converter/StopPlace.kt
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlText
+import java.math.BigDecimal
 
 @JacksonXmlRootElement(localName = "StopPlace")
 class StopPlace(
@@ -56,8 +57,8 @@ class StopPlace(
     }
 
     class Location {
-        var longitude: Double? = null
-        var latitude: Double? = null
+        var longitude: BigDecimal? = null
+        var latitude: BigDecimal? = null
     }
 
     class AccessibilityAssessment {
